The first commit for AIPS!
Archenoth IPS, or "AIPS" is just going to be a small, but ridiculously
portable IPS patcher with binaries that can be built for a wide variety
of machines using only the standard C libraries, the goal is to make a
reliable backend application that anyone can use for either patching
files, creating patches, or using it to extend arbirtary applications
and programs. And eventually, after I have everything working up to
spec, to possibly make statically compiled libraries to make the
handling of binary patches simplistic for anything that can use a C
library.

Initial commit to keep up with my progress working on AIPS since it is a
small side-project I have very little time for...

In this commit:
- A well-formed Makefile with directives for a lot, including an Emacs
  flymake rule. (Makefile)
- The start of my C code, it reads arguments, reads a few bytes, some
  some error checking, but other then that, not a lot. (Work has started
  on reading arbitrarily-sized records) (AIPS.c)
- Definitions and struct initializations, as well as very minimal macros
  for bitshifting digits in the header file (AIPS.h)...